+++ resources/assets/js/app.js
@@ -2,13 +2,19 @@ import { createApp } from 'vue'
 import Chart from "./components/Chart.vue";
 import BarChart from "./components/BarChart.vue";
 import VerticalChart from "./components/VerticalChart.vue";
+import Search from "./components/Search.vue";
+import axios from 'axios';
+import VueAxios from 'vue-axios';
 
 const eva = require('eva-icons');
 
-createApp({})
+const app = createApp({})
+app.use(VueAxios, axios)
+    .provide('axios', app.config.globalProperties.axios)
     .component('chart', Chart)
     .component('bar-chart', BarChart)
     .component('vertical-chart', VerticalChart)
+    .component('search', Search)
     .mount('#app');
 
 function search() {