+++ routes/web.php
@@ -19,7 +19,7 @@ Route::get('/ranking/author/{sort?}', 'RankingController@getIndexAuthor')->where
 Route::get('/post/{id}', 'PostController@getIndex');
 Route::get('/player/{id}', 'PlayerController@getIndex')->name('profile.player');
 Route::get('/author/{name}', 'AuthorController@getIndex');
-Route::get('/search', 'SearchController@postSearch');
+Route::post('/search', 'SearchController@postSearch');
 Route::get('/changelog', 'ChangelogController@getIndex');
 Route::get('/faq', 'FaqController@getIndex');
 Route::get('/stats', 'StatsController@getIndex');